Added more specs; much to do yet
Need to define good default vectors for each metric (instead of hardcoding in,
e.g., the haversine distance specs). Also, need to finish writing the pending
specs -- most of them are straightforward and some are unnecessary.

After the specs are all written out (and passing, of course), must write the
Chebyshev distance, make its specs pass, and it's ok to release 0.0.5.

The next step is writing more distance measures. (...)